git_api_project/requirements.txt
PyGithub==1.44.1
requests==2.22.0
keyring==19.2.0
